Species transport working
Got mass transport working and the model solves with dC_dt active. Also cleaned up the model and moved charge and mass transport into their own files to clean up the code.  There is an active issue where the concentration of water in the anode goes negative as current increases. Likely an issue in the diffusion coefficent function.